requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.3.3
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
certifi==2019.11.28
cffi==1.14.4
channels==3.0.3
chardet==3.0.4
CodeConvert==3.0.2
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.1
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.5
django-admin==2.0.1
django-allauth==0.44.0
django-ckeditor==6.0.0
django-classy-tags==2.0.0
django-cms==3.8.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-formtools==2.2
django-js-asset==1.2.2
django-models-ext==1.1.10
django-sekizai==2.0.0
django-short-url==1.1.6
django-six==1.0.4
django-treebeard==4.4
djangocms-admin-style==2.0.2
djangocms-attributes-field==2.0.0
djangocms-style==3.0.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Faker==5.6.1
filelock==3.0.12
furl==2.1.0
hyperlink==21.0.0
idna==2.8
importlib-metadata==3.4.0
importlib-resources==5.0.0
incremental==17.5.0
isort==4.3.21
jwt==1.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
orderedmultidict==1.0.1
Pillow==8.1.0
pilo==0.6.2
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
requests==2.22.0
requests-oauthlib==1.3.0
screen==1.0.1
service-identity==18.1.0
shortuuid==1.0.1
six==1.14.0
sqlparse==0.4.1
sslcommerz-python==0.0.7
text-unidecode==1.3
TimeConvert==1.5.2
Twisted==20.3.0
txaio==20.12.1
typed-ast==1.4.0
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.7
virtualenv==20.3.1
wrapt==1.11.2
xlwt==1.3.0
zipp==3.1.0
zope.interface==5.2.0
